Add modules "CellState" & "InitializeBoard", but some errors are left
