+++ src/modules/CellState.ts
@@ -1,0 +1,16 @@
+const UNOPENED_CELL = "cell cell--unopened";
+const OPENED_CELL = "cell cell--opened";
+const FLAGGED_CELL = "cell cell--unopened cell--flagged";
+const WRONGLY_FLAGGED_CELL =
+  "cell cell--unopened cell--flagged cell--flagged-wrongly";
+const MINED_CELL = "cell cell--unopened cell--mined";
+const EXPLODED_CELL = "cell cell--exploded";
+
+export {
+  UNOPENED_CELL,
+  OPENED_CELL,
+  FLAGGED_CELL,
+  WRONGLY_FLAGGED_CELL,
+  MINED_CELL,
+  EXPLODED_CELL,
+};
